Revert "chore: refactor `anvil` for library use" (#10135)
Revert "chore: refactor `anvil` for library use (#10130)"

This reverts commit 21bf35b62948741719622b1abde22768d24238f8.
