--- crates/anvil/bin/main.rs
@@ -1,14 +1,0 @@
-//! The `anvil` CLI: a fast local Ethereum development node, akin to Hardhat Network, Tenderly.
-
-use anvil::args::run;
-
-#[cfg(all(feature = "jemalloc", unix))]
-#[global_allocator]
-static ALLOC: tikv_jemallocator::Jemalloc = tikv_jemallocator::Jemalloc;
-
-fn main() {
-    if let Err(err) = run() {
-        let _ = foundry_common::sh_err!("{err:?}");
-        std::process::exit(1);
-    }
-}
