+++ crates/forge/bin/main.rs
@@ -1,6 +1,3 @@
-//! The `forge` CLI: build, test, fuzz, debug and deploy Solidity contracts, like Hardhat, Brownie,
-//! Ape.
-
 use forge::args::run;
 
 #[cfg(all(feature = "jemalloc", unix))]
